% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT data}
\format{
A data frame with 144 rows and 7 columns:
\describe{
  \item{RIVER}{from which river the fish was found}
  \item{MILE}{in which mile of the river the fish was found}
  \item{SPECIES}{what species the fish is}
  \item{LENGTH}{length of the fish}
  \item{WEIGHT}{weight of the fish}
  \item{DDT}{ddt value found in the fish}
  ...
}
}
\source{
<Statistics for Engineering and the Sciences Textbook>
}
\usage{
ddt
}
\description{
A data set that contains data of different species of fish
}
\keyword{datasets}
