% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire data}
\format{
A data frame with 15 rows and 2 columns:
\describe{
  \item{DAMAGE}{Distance the fire travelled}
  \item{DISTANCE}{Level of damage the fire dealt}
  ...
}
}
\source{
<Statistics for Engineering and the Sciences Textbook>
}
\usage{
fire
}
\description{
A data set that gives the distance and damage of each recorded fire
}
\keyword{datasets}
