% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mode.R
\name{getmode}
\alias{getmode}
\title{Title}
\usage{
getmode(v)
}
\arguments{
\item{v}{A quantitative vector of data}
}
\value{
the mode of the given quantitative vector of data
}
\description{
Title
}
\examples{
getmode(c(1, 1, 2 , 3))
}
