% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Title}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{how many iterations of the n binomial trials should be completed}

\item{n}{the number binomial trials}

\item{p}{the probability of a success}
}
\value{
iter plots of the binomial trial
}
\description{
Title
}
\examples{
mybin(iter = 100, n = 10, p = 0.5)
}
