% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations the sample should be taken}

\item{x}{the vector of data to perform the function on}

\item{fun}{the function to use on the given data}

\item{alpha}{the confidence level desired of the confidence interval}

\item{cx}{point size on the graph}

\item{...}{any other parameter}
}
\value{
a graph of the given population value parameter along with a confidence interval and point estimate value
}
\description{
myboot2
}
\examples{
myboot2(x=c(3,4,5),iter=10000,fun="mean",alpha=0.5)
}
