% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{Title mycltu}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}

\item{a}{beginning of sample}

\item{b}{end of sample}
}
\value{
A plot including the histogram and a theoretical curve of the normal distribution
}
\description{
Title mycltu
}
\examples{
mycltu(n=10,iter=100)
}
