% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{the logarathmic function to apply to the x and param parameters}

\item{x}{the set of data}

\item{param}{the sequence over which to evaluate}

\item{...}{other graphical specifications}
}
\value{
a graph of the maximum likelihood given the parameters
}
\description{
mymaxlik
}
\examples{
logbin=function(x,param) log(dbinom(x,prob=param,size=10))
mymaxlik(x=c(4,6,7,6,5),param=seq(0,1,length=1000),lfun=logbin)
}
