% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Title}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{the mean of the normal distribution}

\item{sigma}{the standard deviation of the normal distribution}

\item{a}{the upper limit of the sequence of which the area is being found}
}
\value{
a plot of the normal distribution given the parameters
}
\description{
Title
}
\examples{
myncurve(mu = 10, sigma = 5, a = 6)
}
