% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Title}
\usage{
ntickets(N = 200, gamma = 0.02, p = 0.95)
}
\arguments{
\item{N}{the minimun number of tickets to be sold}

\item{gamma}{the probability that the flight is overbooked}

\item{p}{the probability that a passenger shows up}
}
\value{
a named list containing nd, nc, N, p and gamma and two plots, one of the discrete value and one of the continuous approximation
}
\description{
Title
}
\examples{
ntickets(N = 400, gamma = 0.02, p = 0.95)

}
